Migrate HighlightListElement to plain CSS
Migrates the HighlightListElement component from styled-components to plain
CSS while maintaining backward compatibility.

Changes:
- HighlightListElement.new.tsx: Plain CSS/React implementation with:
  - HighlightOuterWrapper: Simple div with CSS class
  - HighlightContentWrapper: Div with dynamic CSS variables for highlight colors
  - HiddenLabel: Accessibility label using plain CSS

- HighlightListElement.css: Plain CSS styles including:
  - Wrapper borders and print styles
  - Dynamic color handling via CSS variables (--highlight-focused-color,
    --highlight-passive-color)
  - ContentExcerpt styling via child selectors

- HighlightListElement.legacy.tsx: Styled-components wrappers for backward
  compatibility

- HighlightListElement.tsx: Re-export shim maintaining original import path

Key Pattern:
The component uses CSS variables to handle dynamic color-based styling. The
HighlightContentWrapper component accepts a `color` prop, looks up the
corresponding highlight style, and sets CSS variables that are used by the
CSS file.

Related: CORE-2282


Update HighlightListElement.new.tsx
